Lots of changes, but most of the changes are Compiler Warning fixes. The important fix is DeOptimizeCodeA in vip.cpp, which keeps the VC++ 2010 optimization routine from killing the effect of some code. Some debugging-friendly additions have been made to vip_draw.inc, which allow you to put it into vip.cpp to debug, and then lets you put it back into the .inc file. This serves no purpose other than to keep vip.cpp from getting cluttered. Overall, a rather big project in terms of time invested.

+++ trunk/dd.cpp
@@ -301,7 +301,7 @@ void render() {
 		pcejin.height = espec.DisplayRect.h;
 		CreateDDrawBuffers();
 		if(!pcejin.maximized)
-			ScaleScreen(pcejin.windowSize);
+			ScaleScreen((float)pcejin.windowSize);
 	}
 
 	pcejin.width = espec.DisplayRect.w;
